+++ Models/Entities/Order.cs
@@ -1,16 +1,17 @@
-﻿using Models.Entities.Enums;
+﻿using Microsoft.EntityFrameworkCore;
+using Models.Entities.Enums;
 using System.ComponentModel.DataAnnotations;
 
 namespace Models.Entities;
 
 /// <summary>
-/// Represents a customer order, including its unique identifier, associated
-/// customer, and current status.
+/// Represents a order, including its unique identifier, associated
+/// user, and current status.
 /// </summary>
 /// <remarks>
 /// The <see cref="Order"/> class encapsulates information about a single order
-/// placed by a customer. It includes references to the customer, the order's status,
-/// and unique identifiers for both the order and the customer.
+/// placed by a user. It includes references to the user, the order's status,
+/// and unique identifiers for both the order and the user.
 /// </remarks>
 public class Order : BaseEntity {
 	/// <summary>
@@ -23,18 +24,19 @@ public class Order : BaseEntity {
 	}
 
 	/// <summary>
-	/// Gets or sets the foreign key of the customer associated with this entity.
+	/// Gets or sets the foreign key of the user associated with this entity.
 	/// </summary>
-	[Display(Name = "CustomerId")]
-	public required long CustomerId {
+	[Display(Name = "CoffeeUserId")]
+	public required long CoffeeUserId {
 		get; set;
 	}
 
 	/// <summary>
-	/// Gets or sets the customer associated with this entity.
+	/// Gets or sets the user associated with this entity.
 	/// </summary>
-	[Display(Name = "Customer")]
-	public Customer? Customer {
+	[DeleteBehavior(DeleteBehavior.Restrict)]
+	[Display(Name = "CoffeeUser")]
+	public CoffeeUser? CoffeeUser {
 		get; set;
 	}
 
@@ -49,21 +51,21 @@ public class Order : BaseEntity {
 	/// <summary>
 	/// Generates a fixed set of <see cref="Order"/> entities for seeding purposes.
 	/// </summary>
-	/// <param name="customerIds">
-	/// An array of customer identifiers. Identifiers are pseudo-randomly mapped to one or more
-	/// generated <see cref="PaymentDetail"/> instance.
+	/// <param name="userIds">
+	/// An array of user identifiers. Identifiers are pseudo-randomly mapped to one or more
+	/// generated <see cref="PaymentDetail"/> instances.
 	/// </param>
 	/// <returns>
 	/// An array of <see cref="Order"/> objects containing predefined order data
-	/// associated with the provided customer identifiers.
+	/// associated with the provided user identifiers.
 	/// </returns>
-	public static Order[] SeedingData(long[] customerIds) {
+	public static Order[] SeedingData(long[] userIds) {
 		Random rnd = new();
 		var orders = new Order[20];
 
 		for (int i = 0; i < orders.Length; i++) {
 			orders[i] = new() {
-				CustomerId = customerIds[rnd.Next(customerIds.Length)],
+				CoffeeUserId = userIds[rnd.Next(userIds.Length)],
 			};
 		}